+++ audio_be/src/commonMain/kotlin/filters/NoOpAudioFilter.kt
@@ -4,5 +4,5 @@ package io.peekandpoke.klang.audio_be.filters
  * No-op filter
  */
 object NoOpAudioFilter : AudioFilter {
-    override fun process(buffer: DoubleArray, offset: Int, length: Int) {}
+    override fun process(buffer: FloatArray, offset: Int, length: Int) {}
 }
