+++ audio_be/src/commonTest/kotlin/effects/DuckingSpec.kt
@@ -17,15 +17,15 @@ class DuckingSpec : StringSpec({
         )
 
         // Create input signal (constant level)
-        val input = DoubleArray(100) { 1.0 }
+        val input = FloatArray(100) { 1.0f }
 
         // Create sidechain signal (loud trigger)
-        val sidechain = DoubleArray(100) { 0.8 }
+        val sidechain = FloatArray(100) { 0.8f }
 
         ducking.process(input, sidechain, 100)
 
         // Input should be reduced significantly
-        val avgLevel = input.map { abs(it) }.average()
+        val avgLevel = input.map { abs(it.toDouble()) }.average()
         avgLevel shouldBeLessThan 0.5
     }
 
@@ -37,24 +37,24 @@ class DuckingSpec : StringSpec({
         )
 
         // Process with active sidechain
-        val input1 = DoubleArray(100) { 1.0 }
-        val sidechain1 = DoubleArray(100) { 1.0 }
+        val input1 = FloatArray(100) { 1.0f }
+        val sidechain1 = FloatArray(100) { 1.0f }
         ducking.process(input1, sidechain1, 100)
 
         // Process with silent sidechain (should return to normal)
-        val input2 = DoubleArray(1000) { 1.0 }
-        val sidechain2 = DoubleArray(1000) { 0.0 }
+        val input2 = FloatArray(1000) { 1.0f }
+        val sidechain2 = FloatArray(1000) { 0.0f }
         ducking.process(input2, sidechain2, 1000)
 
         // Should return close to full volume
-        val endLevel = input2.takeLast(100).map { abs(it) }.average()
+        val endLevel = input2.takeLast(100).map { abs(it.toDouble()) }.average()
         endLevel shouldBe (1.0 plusOrMinus 0.1)
     }
 
     "Depth parameter controls ducking amount" {
-        val input1 = DoubleArray(100) { 1.0 }
-        val input2 = DoubleArray(100) { 1.0 }
-        val sidechain = DoubleArray(100) { 1.0 }
+        val input1 = FloatArray(100) { 1.0f }
+        val input2 = FloatArray(100) { 1.0f }
+        val sidechain = FloatArray(100) { 1.0f }
 
         val duckingLight = Ducking(sampleRate, 0.01, depth = 0.3)
         val duckingHeavy = Ducking(sampleRate, 0.01, depth = 0.9)
@@ -62,8 +62,8 @@ class DuckingSpec : StringSpec({
         duckingLight.process(input1, sidechain, 100)
         duckingHeavy.process(input2, sidechain, 100)
 
-        val avgLight = input1.map { abs(it) }.average()
-        val avgHeavy = input2.map { abs(it) }.average()
+        val avgLight = input1.map { abs(it.toDouble()) }.average()
+        val avgHeavy = input2.map { abs(it.toDouble()) }.average()
 
         // Heavy ducking should reduce more
         avgHeavy shouldBeLessThan avgLight
@@ -73,19 +73,19 @@ class DuckingSpec : StringSpec({
         val ducking = Ducking(sampleRate, 0.01, 1.0)
 
         // Duck the signal
-        val input = DoubleArray(100) { 1.0 }
-        val sidechain = DoubleArray(100) { 1.0 }
+        val input = FloatArray(100) { 1.0f }
+        val sidechain = FloatArray(100) { 1.0f }
         ducking.process(input, sidechain, 100)
 
         // Reset
         ducking.reset()
 
         // Should process at full volume immediately
-        val input2 = DoubleArray(10) { 1.0 }
-        val sidechain2 = DoubleArray(10) { 0.0 }
+        val input2 = FloatArray(10) { 1.0f }
+        val sidechain2 = FloatArray(10) { 0.0f }
         ducking.process(input2, sidechain2, 10)
 
-        input2[0] shouldBe (1.0 plusOrMinus 0.01)
+        input2[0].toDouble() shouldBe (1.0 plusOrMinus 0.01)
     }
 
     "Ducking with zero depth has no effect" {
@@ -95,13 +95,13 @@ class DuckingSpec : StringSpec({
             depth = 0.0
         )
 
-        val input = DoubleArray(100) { 1.0 }
-        val sidechain = DoubleArray(100) { 1.0 }
+        val input = FloatArray(100) { 1.0f }
+        val sidechain = FloatArray(100) { 1.0f }
 
         ducking.process(input, sidechain, 100)
 
         // Should remain at full volume
-        val avgLevel = input.map { abs(it) }.average()
+        val avgLevel = input.map { abs(it.toDouble()) }.average()
         avgLevel shouldBe (1.0 plusOrMinus 0.01)
     }
 
@@ -109,34 +109,34 @@ class DuckingSpec : StringSpec({
         val ducking = Ducking(sampleRate, 0.01, 0.8)
 
         // Left channel
-        val inputLeft = DoubleArray(100) { 1.0 }
-        val sidechainLeft = DoubleArray(100) { 0.8 }
+        val inputLeft = FloatArray(100) { 1.0f }
+        val sidechainLeft = FloatArray(100) { 0.8f }
 
         // Right channel
-        val inputRight = DoubleArray(100) { 1.0 }
-        val sidechainRight = DoubleArray(100) { 0.8 }
+        val inputRight = FloatArray(100) { 1.0f }
+        val sidechainRight = FloatArray(100) { 0.8f }
 
         ducking.process(inputLeft, sidechainLeft, 100)
         ducking.process(inputRight, sidechainRight, 100)
 
         // Both channels should be ducked
-        val avgLeft = inputLeft.map { abs(it) }.average()
-        val avgRight = inputRight.map { abs(it) }.average()
+        val avgLeft = inputLeft.map { abs(it.toDouble()) }.average()
+        val avgRight = inputRight.map { abs(it.toDouble()) }.average()
 
         avgLeft shouldBeLessThan 0.5
         avgRight shouldBeLessThan 0.5
     }
 
     "Attack time affects return speed" {
-        val input1 = DoubleArray(500) { 1.0 }
-        val input2 = DoubleArray(500) { 1.0 }
+        val input1 = FloatArray(500) { 1.0f }
+        val input2 = FloatArray(500) { 1.0f }
 
         val duckingFast = Ducking(sampleRate, attackSeconds = 0.001, depth = 0.8)
         val duckingSlow = Ducking(sampleRate, attackSeconds = 0.1, depth = 0.8)
 
         // Duck both with active sidechain for first 100 samples
-        val sidechainActive = DoubleArray(100) { 1.0 }
-        val sidechainSilent = DoubleArray(400) { 0.0 }
+        val sidechainActive = FloatArray(100) { 1.0f }
+        val sidechainSilent = FloatArray(400) { 0.0f }
 
         duckingFast.process(input1.sliceArray(0 until 100), sidechainActive, 100)
         duckingFast.process(input1.sliceArray(100 until 500), sidechainSilent, 400)
@@ -145,8 +145,8 @@ class DuckingSpec : StringSpec({
         duckingSlow.process(input2.sliceArray(100 until 500), sidechainSilent, 400)
 
         // Fast should recover more by sample 200
-        val fastLevel = input1.sliceArray(150 until 200).map { abs(it) }.average()
-        val slowLevel = input2.sliceArray(150 until 200).map { abs(it) }.average()
+        val fastLevel = input1.sliceArray(150 until 200).map { abs(it.toDouble()) }.average()
+        val slowLevel = input2.sliceArray(150 until 200).map { abs(it.toDouble()) }.average()
 
         fastLevel shouldBe (slowLevel plusOrMinus 0.1)
         fastLevel shouldBeLessThan (slowLevel + 0.2) // Fast returns faster
